FaceRecognition-TripletLoss/myenv/requirements.txt
facenet-pytorch
torch
torchvision
numpy
opencv-python
Pillow
